+++ symphony-bdk-core/src/main/java/com/symphony/bdk/core/service/message/MessageService.java
@@ -1,10 +1,18 @@
 package com.symphony.bdk.core.service.message;
 
+import static java.util.Collections.emptyList;
+import static java.util.Collections.emptyMap;
+
 import com.symphony.bdk.core.auth.AuthSession;
+import com.symphony.bdk.core.retry.RetryWithRecovery;
 import com.symphony.bdk.core.retry.RetryWithRecoveryBuilder;
+import com.symphony.bdk.core.service.OboService;
+import com.symphony.bdk.core.service.message.model.Attachment;
+import com.symphony.bdk.core.service.message.model.Message;
 import com.symphony.bdk.core.service.pagination.PaginatedApi;
 import com.symphony.bdk.core.service.pagination.PaginatedService;
 import com.symphony.bdk.core.service.stream.constant.AttachmentSort;
+import com.symphony.bdk.core.util.function.SupplierWithApiException;
 import com.symphony.bdk.gen.api.AttachmentsApi;
 import com.symphony.bdk.gen.api.DefaultApi;
 import com.symphony.bdk.gen.api.MessageApi;
@@ -22,14 +30,21 @@ import com.symphony.bdk.gen.api.model.V4ImportResponse;
 import com.symphony.bdk.gen.api.model.V4ImportedMessage;
 import com.symphony.bdk.gen.api.model.V4Message;
 import com.symphony.bdk.gen.api.model.V4Stream;
+import com.symphony.bdk.http.api.ApiClient;
+import com.symphony.bdk.http.api.ApiClientBodyPart;
+import com.symphony.bdk.http.api.ApiException;
 import com.symphony.bdk.http.api.util.ApiUtils;
+import com.symphony.bdk.http.api.util.TypeReference;
 import com.symphony.bdk.template.api.TemplateEngine;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apiguardian.api.API;
 
+import java.io.File;
 import java.time.Instant;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Stream;
 
 import javax.annotation.Nonnull;
@@ -41,14 +56,18 @@ import javax.annotation.Nonnull;
  */
 @Slf4j
 @API(status = API.Status.STABLE)
-public class MessageService extends OboMessageService {
+public class MessageService implements OboMessageService, OboService<OboMessageService> {
 
+  private final MessagesApi messagesApi;
   private final MessageApi messageApi;
   private final MessageSuppressionApi messageSuppressionApi;
   private final StreamsApi streamsApi;
   private final PodApi podApi;
   private final AttachmentsApi attachmentsApi;
   private final DefaultApi defaultApi;
+  private final AuthSession authSession;
+  private final TemplateEngine templateEngine;
+  private final RetryWithRecoveryBuilder<?> retryBuilder;
 
   public MessageService(
       final MessagesApi messagesApi,
@@ -62,13 +81,30 @@ public class MessageService extends OboMessageService {
       final TemplateEngine templateEngine,
       final RetryWithRecoveryBuilder<?> retryBuilder
   ) {
-    super(messagesApi, authSession, templateEngine, retryBuilder);
+    this.messagesApi = messagesApi;
     this.messageApi = messageApi;
     this.messageSuppressionApi = messageSuppressionApi;
     this.streamsApi = streamsApi;
     this.podApi = podApi;
     this.attachmentsApi = attachmentsApi;
+    this.authSession = authSession;
+    this.templateEngine = templateEngine;
     this.defaultApi = defaultApi;
+    this.retryBuilder = retryBuilder;
+  }
+
+  @Override
+  public OboMessageService obo(AuthSession oboSession) {
+    return new MessageService(messagesApi, messageApi, messageSuppressionApi, streamsApi, podApi, attachmentsApi,
+        defaultApi, oboSession, templateEngine, retryBuilder);
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
+  public TemplateEngine templates() {
+    return this.templateEngine;
   }
 
   /**
@@ -133,6 +169,78 @@ public class MessageService extends OboMessageService {
     final int actualTotalSize = totalSize == null ? 50 : totalSize;
 
     return new PaginatedService<>(api, actualChunkSize, actualTotalSize).stream();
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
+  public V4Message send(@Nonnull V4Stream stream, @Nonnull String message) {
+    return send(stream.getStreamId(), message);
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
+  public V4Message send(@Nonnull String streamId, @Nonnull String message) {
+    return this.send(streamId, Message.builder().content(message).build());
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
+  public V4Message send(@Nonnull V4Stream stream, @Nonnull Message message) {
+    return this.send(stream.getStreamId(), message);
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
+  public V4Message send(@Nonnull String streamId, @Nonnull Message message) {
+    return this.executeAndRetry("send", () ->
+        this.doSend(streamId, message, this.authSession.getSessionToken(), this.authSession.getKeyManagerToken())
+    );
+  }
+
+  /**
+   * The generated {@link MessagesApi#v4StreamSidMessageCreatePost(String, String, String, String, String, String, File, File)}
+   * does not allow to send multiple attachments as well as in-memory files, so we have to "manually" process this call.
+   */
+  private V4Message doSend(String streamId, Message message, String sessionToken, String keyManagerToken) throws ApiException {
+    final ApiClient apiClient = this.messagesApi.getApiClient();
+    final Map<String, Object> form = new HashMap<>();
+    form.put("message", message.getContent());
+    form.put("data", message.getData());
+    form.put("version", message.getVersion());
+    form.put("attachment", toApiClientBodyParts(message.getAttachments()));
+    form.put("preview", toApiClientBodyParts(message.getPreviews()));
+
+    final Map<String, String> headers = new HashMap<>();
+    headers.put("sessionToken", apiClient.parameterToString(sessionToken));
+    headers.put("keyManagerToken", apiClient.parameterToString(keyManagerToken));
+
+    return apiClient.invokeAPI(
+        "/v4/stream/" + apiClient.escapeString(streamId) + "/message/create",
+        "POST",
+        emptyList(),
+        null, // for 'multipart/form-data', body can be null
+        headers,
+        emptyMap(),
+        form,
+        apiClient.selectHeaderAccept("application/json"),
+        apiClient.selectHeaderContentType("multipart/form-data"),
+        new String[0],
+        new TypeReference<V4Message>() {}
+    ).getData();
+  }
+
+  private static ApiClientBodyPart[] toApiClientBodyParts(List<Attachment> attachments) {
+    return attachments.stream()
+        .map(a -> new ApiClientBodyPart(a.getContent(), a.getFilename()))
+        .toArray(ApiClientBodyPart[]::new);
   }
 
   /**
@@ -298,5 +406,9 @@ public class MessageService extends OboMessageService {
 
   private static Long getEpochMillis(Instant instant) {
     return instant == null ? null : instant.toEpochMilli();
+  }
+
+  private <T> T executeAndRetry(String name, SupplierWithApiException<T> supplier) {
+    return RetryWithRecovery.executeAndRetry(retryBuilder, name, supplier);
   }
 }
